Implement rebase changes
