+++ systemtests/src/test/java/io/enmasse/systemtest/iot/isolated/project/ManagedTest.java
@@ -2,18 +2,33 @@
  * Copyright 2019, EnMasse authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.enmasse.systemtest.iot.project;
+package io.enmasse.systemtest.iot.isolated.project;
 
 import static io.enmasse.systemtest.utils.AddressSpaceUtils.addressSpaceExists;
 import static io.enmasse.systemtest.utils.TestUtils.waitUntilConditionOrFail;
 import static java.time.Duration.ofMinutes;
 import static java.time.Duration.ofSeconds;
 
+import java.nio.ByteBuffer;
 import java.util.function.BiConsumer;
+
+import io.enmasse.iot.model.v1.IoTConfig;
+import io.enmasse.iot.model.v1.IoTConfigBuilder;
+import io.enmasse.systemtest.Endpoint;
+import io.enmasse.systemtest.bases.TestBase;
+import io.enmasse.systemtest.bases.iot.ITestIoTIsolated;
+import io.enmasse.systemtest.certs.CertBundle;
+import io.enmasse.systemtest.iot.CredentialsRegistryClient;
+import io.enmasse.systemtest.iot.DefaultDeviceRegistry;
+import io.enmasse.systemtest.iot.DeviceRegistryClient;
+import io.enmasse.systemtest.platform.apps.SystemtestsKubernetesApps;
+import io.enmasse.systemtest.utils.CertificateUtils;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtensionContext;
 import org.slf4j.Logger;
 
 import io.enmasse.address.model.Address;
@@ -36,8 +51,7 @@ import io.enmasse.user.model.v1.UserList;
 import io.fabric8.kubernetes.client.dsl.MixedOperation;
 import io.fabric8.kubernetes.client.dsl.Resource;
 
-@Tag(TestTag.SHARED_IOT)
-public class ManagedTest extends AbstractIoTProjectTestBase  {
+public class ManagedTest extends TestBase implements ITestIoTIsolated {
 
     private static final Logger log = CustomLogger.getLogger();
 
@@ -45,20 +59,58 @@ public class ManagedTest extends AbstractIoTProjectTestBase  {
     private MixedOperation<Address, AddressList, DoneableAddress, Resource<Address, DoneableAddress>> addressClient;
     private MixedOperation<AddressSpace, AddressSpaceList, DoneableAddressSpace, Resource<AddressSpace, DoneableAddressSpace>> addressSpaceClient;
     private MixedOperation<User, UserList, DoneableUser, Resource<User, DoneableUser>> userClient;
+    protected DeviceRegistryClient registryClient;
+    protected CredentialsRegistryClient credentialsClient;
 
     @BeforeEach
-    public void initClients () {
+    public void initClients () throws Exception {
+        CertBundle certBundle = CertificateUtils.createCertBundle();
+        IoTConfig iotConfig = new IoTConfigBuilder()
+                .withNewMetadata()
+                .withName("default")
+                .endMetadata()
+                .withNewSpec()
+                .withNewServices()
+                .withDeviceRegistry(DefaultDeviceRegistry.newInfinispanBased())
+                .endServices()
+                .withNewAdapters()
+                .withNewMqtt()
+                .withNewEndpoint()
+                .withNewKeyCertificateStrategy()
+                .withCertificate(ByteBuffer.wrap(certBundle.getCert().getBytes()))
+                .withKey(ByteBuffer.wrap(certBundle.getKey().getBytes()))
+                .endKeyCertificateStrategy()
+                .endEndpoint()
+                .endMqtt()
+                .endAdapters()
+                .endSpec()
+                .build();
+
+        isolatedIoTManager.createIoTConfig(iotConfig);
+
+        final Endpoint deviceRegistryEndpoint = kubernetes.getExternalEndpoint("device-registry");
+        registryClient = new DeviceRegistryClient(kubernetes, deviceRegistryEndpoint);
+        credentialsClient = new CredentialsRegistryClient(kubernetes, deviceRegistryEndpoint);
         this.client = kubernetes.getIoTProjectClient(this.iotProjectNamespace);
         this.addressClient = kubernetes.getAddressClient(this.iotProjectNamespace);
         this.addressSpaceClient = kubernetes.getAddressSpaceClient(this.iotProjectNamespace);
         this.userClient = kubernetes.getUserClient(this.iotProjectNamespace);
     }
 
+    @AfterEach
+    void cleanEnv(ExtensionContext context) throws Exception {
+        if (context.getExecutionException().isPresent()) { //test failed
+            logCollector.collectHttpAdapterQdrProxyState();
+        }
+        SystemtestsKubernetesApps.deleteInfinispanServer(kubernetes.getInfraNamespace());
+    }
+
     @Test
     public void testChangeAddressSpace() throws Exception {
 
-        var project = IoTUtils.getBasicIoTProjectObject("iot1", "as1", this.iotProjectNamespace);
-        createIoTProject(project);
+        var project = IoTUtils.getBasicIoTProjectObject("iot1", "as1",
+                this.iotProjectNamespace, getDefaultAddressSpacePlan());
+        isolatedIoTManager.createIoTProject(project);
 
         assertManagedResources(Assertions::assertNotNull, project, "as1");
